requirements.txt
faiss_cpu==1.12.0
langchain_community==0.4.1
langchain_text_splitters==1.0.0
python-dotenv==1.2.1
sentence_transformers==5.1.2
tavily_python==0.7.12